test: dedupe child-process test runner, harden happy-path assertions
Two cubic review findings on the new tests/lib/validator-cli.test.js
(EGC-539 PR #1151):

1. Its child-process runner duplicated the temp-file + execFileSync +
   error-code + finally-unlink scaffolding already in
   tests/ci/validators.test.js's runSourceViaTempFile(). Extracted both
   into tests/lib/run-source-in-child-process.js; both suites now share
   one implementation instead of drifting independently on timeout/env/
   cleanup behavior.

2. The skipIfMissing/finishValidation happy-path cases (path exists,
   hasErrors false) ran in the test process itself even though both
   helpers call process.exit() on other branches. If either ever
   regressed into exiting on that path too, it would silently kill the
   whole suite mid-run (both helpers exit 0 on their "nothing wrong"
   path, so that reads as success with fewer tests, not a failure) instead
   of failing loudly. Added child-process counterparts that assert code
   after the call still runs, so a stray exit shows up as a missing
   REACHED/UNREACHABLE marker instead of a shorter test count.
